fix: extract GraphQLShemaRegistration interface
