+++ graphql-jpa-query-autoconfigure/src/main/java/com/introproventures/graphql/jpa/query/autoconfigure/GraphQLSchemaAutoConfiguration.java
@@ -3,14 +3,15 @@ package com.introproventures.graphql.jpa.query.autoconfigure;
 import java.util.ArrayList;
 import java.util.List;
 
-import graphql.GraphQL;
-import graphql.schema.GraphQLSchema;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.util.CollectionUtils;
+
+import graphql.GraphQL;
+import graphql.schema.GraphQLSchema;
 
 @Configuration
 @ConditionalOnClass(GraphQL.class)
@@ -28,7 +29,7 @@ public class GraphQLSchemaAutoConfiguration {
     @Bean
     @ConditionalOnMissingBean(GraphQLSchema.class)
     public GraphQLSchemaFactoryBean graphQLSchemaFactoryBean() {
-    	GraphQLShemaRegistration graphQLShemaRegistration = new GraphQLShemaRegistration();
+    	GraphQLShemaRegistrationImpl graphQLShemaRegistration = new GraphQLShemaRegistrationImpl();
 
         for (GraphQLSchemaConfigurer configurer : graphQLSchemaConfigurers) {
             configurer.configure(graphQLShemaRegistration);
