+++ graphql-jpa-query-autoconfigure/src/main/java/com/introproventures/graphql/jpa/query/autoconfigure/GraphQLShemaRegistration.java
@@ -1,20 +1,10 @@
 package com.introproventures.graphql.jpa.query.autoconfigure;
 
-import java.util.LinkedHashSet;
-import java.util.Set;
-
 import graphql.schema.GraphQLSchema;
 
-public class GraphQLShemaRegistration {
+public interface GraphQLShemaRegistration {
 
-	Set<GraphQLSchema> managedGraphQLSchemas = new LinkedHashSet<GraphQLSchema>();
+	public void register(GraphQLSchema graphQLSchema);
 
-	public void register(GraphQLSchema graphQLSchema) {
-		managedGraphQLSchemas.add(graphQLSchema);
-	}
-
-	public GraphQLSchema[] getManagedGraphQLSchemas() {
-		return managedGraphQLSchemas.toArray(new GraphQLSchema[] {});
-	}
 
 }
